Refactor MaterializedViewRewriter to avoid creating multi times for multi MVs
