+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/LineageFactory.java
@@ -62,4 +62,10 @@ public class LineageFactory {
             return null;
         }
     }
+
+    public static Map<ColumnRefOperator, ScalarOperator> getLineage(OptExpression expression,
+                                                                    ColumnRefFactory refFactory) {
+        LineageFactory factory = new LineageFactory(expression, refFactory);
+        return factory.getLineage();
+    }
 }
